stock-ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6

# LangGraph and LangChain dependencies
langgraph>=0.2.27
langchain>=0.2.16
langchain-core>=0.2.39
langchain-openai>=0.1.23
langchain-community>=0.2.16
openai>=1.51.0

# Database dependencies
sqlalchemy>=2.0.23
asyncpg>=0.28.0
psycopg2-binary>=2.9.7
alembic>=1.12.1

# Additional AI/ML dependencies
tiktoken>=0.7.0
tenacity>=8.2.3
httpx>=0.25.0
aiohttp>=3.9.1

# Financial data dependencies
yfinance>=0.2.18
requests>=2.31.0